# openapi.yaml
openapi: 3.1.0
info:
  title: 알바 관리 서비스 API (근로계약서 / 대타·추가근무 Agent)
  version: 0.1.0
  description: |
    해커톤 MVP용 API 명세.
    - 근로계약서 Agent: 문서 업로드 파싱(UF1, PDF/DOCX/TXT), 줄글 입력 생성(UF2)
    - 대타/추가근무 Agent: 증빙 → VLM 해석 → 온톨로지 매핑 → 급여 재계산
    비동기 작업(파싱·VLM 해석)은 202 + jobId 로 응답하며 폴링/SSE 로 결과를 수신한다.
servers:
  - url: http://localhost:8000/api/v1
    description: 로컬 개발
...
paths:
...
  /contracts/upload:
    post:
      tags: [contracts]
      summary: 근로계약서 문서 업로드 (UF1)
      description: |
        근로계약서 문서(PDF / DOCX / TXT) 업로드 → labor_contracts(status=parsing) 생성
        + contract_parse 잡 트리거. 워커가 문서에서 텍스트를 추출해 계약 조건을 추출한다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [store_id, file]
              properties:
                store_id:
                  type: integer
                  format: int64
                file:
                  type: string
                  format: binary
                  description: 계약서 파일 (application/pdf, .docx, .txt 지원)
      responses:
        '202':
          description: 파싱 잡 접수됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
...
  /contracts/{id}:
    get:
      tags: [contracts]
      summary: 계약 상세 / 추출 결과 조회
      parameters:
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [contracts]
      summary: 추출된 계약 조건 수정 (UF2)
      parameters:
        - $ref: '#/components/parameters/ContractId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTermsInput'
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
  /contracts/{id}/confirm:
    post:
      tags: [contracts]
      summary: 추출값 검증·확정 (UF1)
      description: extracted_terms → contract_terms/work_schedules/ontology_nodes 정형화, status=active
      parameters:
        - $ref: '#/components/parameters/ContractId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTermsInput'
      responses:
        '200':
          description: 확정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
...
  /agent-jobs/{jobId}:
    get:
      tags: [jobs]
      summary: Agent 잡 상태 폴링
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentJob'

  ##################################################################
  # 대타/추가근무 Agent
  ##################################################################
...
components:
  parameters:
    ContractId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    JobAccepted:
      type: object
      properties:
        job_id: { type: integer, format: int64 }
        status: { type: string, example: queued }
        poll_url: { type: string, example: /api/v1/agent-jobs/12 }
    AgentJob:
      type: object
      properties:
        id: { type: integer, format: int64 }
        job_type:
          type: string
          enum: [contract_parse, contract_generate, vlm_interpret]
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        result: { type: object }
        error: { type: string, nullable: true }
    Contract:
      type: object
      properties:
        id: { type: integer, format: int64 }
        store_id: { type: integer, format: int64 }
        employee_id: { type: integer, format: int64, nullable: true }
        source: { type: string, enum: [uploaded, generated] }
        status:
          type: string
          enum: [draft, parsing, extracted, pending_signature, active, expired, failed]
        raw_text: { type: string, nullable: true }
        extracted_terms:
          $ref: '#/components/schemas/ContractTermsInput'
        terms:
          $ref: '#/components/schemas/ContractTermsInput'
        created_at: { type: string, format: date-time }
    ContractTermsInput:
      type: object
      properties:
        hourly_wage: { type: integer, example: 11000 }
        weekly_hours: { type: number, example: 12.0 }
        work_start_date: { type: string, format: date }
        work_end_date: { type: string, format: date, nullable: true }
        pay_day: { type: integer, example: 10 }
        tax_type: { type: string, example: freelance_3_3 }
        weekly_holiday_eligible: { type: boolean }
        schedules:
          type: array
          items:
            type: object
            properties:
              day_of_week: { type: integer, minimum: 0, maximum: 6 }
              start_time: { type: string, example: "18:00" }
              end_time: { type: string, example: "22:00" }
              break_minutes: { type: integer, example: 0 }
...